src/goVulncheck2: rename goVulncheck2.ts to goVulncheck.ts
Old implementation was removed.

Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/455557
Run-TryBot: Hyang-Ah Hana Kim <hyangah@gmail.com>
TryBot-Result: kokoro <noreply+kokoro@google.com>
