+++ src/goVulncheck.ts
